requirements.txt
Flask
nltk
python-docx
pdfminer.six
sentence-transformers
faiss-cpu
langchain
langchain-groq
langchain-community
python-dotenv
